feat (yarzawin-api): user table, authorization with hard-coded user and access token only, update e2e tests, chore: add utility scripts to root package.json, fix (yarzawin-web): docker vite to expose port

+++ apps/yarzawin-api/src/modules/diary/diary.controller.ts
@@ -1,28 +1,31 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common'
+import { Body, Controller, Delete, Get, Param, Post, Put, UseGuards } from '@nestjs/common'
 import { DiaryService } from './diary.service'
-import { CreateDiaryDto, DeleteDiaryDto, UpdateDiaryDto } from './diary.dto'
+import { CreateDiaryDto, UpdateDiaryDto } from './diary.dto'
+import { AuthGuard } from 'src/guards/auth.guard'
+import { Auth } from 'src/decorators/param.decorator'
 
+@UseGuards(AuthGuard)
 @Controller({ path: '/diaries' })
 export class DiaryController {
   constructor(private diaryService: DiaryService) {}
 
   @Get('/')
-  async getDiaries() {
-    return this.diaryService.getDiaryList()
+  async getDiaries(@Auth('userId') userId: string) {
+    return this.diaryService.getDiaryList(userId)
   }
 
   @Post('/')
-  async createDiary(@Body() data: CreateDiaryDto) {
-    return this.diaryService.createDiary(data)
+  async createDiary(@Body() data: CreateDiaryDto, @Auth('userId') userId: string) {
+    return this.diaryService.createDiary({ ...data, userId })
   }
 
   @Put('/:id')
-  async updateDiary(@Body() data: UpdateDiaryDto) {
-    return this.diaryService.updateDiary(data)
+  async updateDiary(@Body() data: UpdateDiaryDto, @Auth('userId') userId: string) {
+    return this.diaryService.updateDiary({ ...data, userId })
   }
 
   @Delete('/:id')
-  async deleteDiary(@Param('id') id: string) {
-    return this.diaryService.deleteDiary({ id })
+  async deleteDiary(@Param('id') id: string, @Auth('userId') userId: string) {
+    return this.diaryService.deleteDiary({ id, userId })
   }
 }
